requirements.txt
kafka-python~=2.2.15
kagglehub~=0.3.12
pandas~=2.3.1
